Extract some Fortran compat functionality from C++ sources
Expose it from the main Python layer instead.
