+++ src/vmecpp/__init__.py
@@ -12,9 +12,8 @@ import numpy as np
 import pydantic
 from beartype import beartype
 
-from vmecpp import _util
+from vmecpp import _util, simsopt_compat
 from vmecpp.cpp import _vmecpp  # bindings to the C++ core
-from vmecpp.cpp.vmecpp import simsopt_compat as _simsopt_compat
 
 
 # This is a pure Python equivalent of VmecINDATAPyWrapper.
@@ -187,7 +186,7 @@ class VmecInput(pydantic.BaseModel):
         with tempfile.TemporaryDirectory() as tmpdir, _util.change_working_directory_to(  # noqa: SIM117
             Path(tmpdir)
         ):  # we call this in a temporary directory because it produces the file in the current working directory
-            with _simsopt_compat.ensure_vmecpp_input(
+            with simsopt_compat.ensure_vmecpp_input(
                 absolute_input_path
             ) as vmecpp_input_file:
                 vmecpp_indata_pywrapper = _vmecpp.VmecINDATAPyWrapper.from_file(
