+++ src/vmecpp/cpp/vmecpp/simsopt_compat/__init__.py
@@ -10,19 +10,11 @@ from vmecpp.cpp.vmecpp.simsopt_compat._indata_to_surfacerzfourier import (
     surfacerzfourier_from_fourier_coeffs,
     surfacerzfourier_from_vmecppindata,
 )
-from vmecpp.cpp.vmecpp.simsopt_compat._vmec_indata_compat import (
-    ensure_vmec2000_input,
-    ensure_vmecpp_input,
-    is_vmec2000_input,
-)
 
 __all__ = [
-    "ensure_vmecpp_input",
-    "ensure_vmec2000_input",
-    "is_vmec2000_input",
-    "surfacerzfourier_from_any_vmec_indata",
-    "surfacerzfourier_from_vmecppindata",
-    "surfacerzfourier_from_fourier_coeffs",
-    "FortranWOutAdapter",
     "VARIABLES_MISSING_FROM_FORTRAN_WOUT_ADAPTER",
+    "FortranWOutAdapter",
+    "surfacerzfourier_from_any_vmec_indata",
+    "surfacerzfourier_from_fourier_coeffs",
+    "surfacerzfourier_from_vmecppindata",
 ]
